Type: Initial production code for Udacity DS python nanodegree.
The code is design to analyse the data of a Bike Share rental Company.The code is to give a complete analytics about it offered services.
The company data is gotten from its various locations found in chicago, new york city and washington, USA.
The code query through the data of any town and for all of them mixed, giving statistics of their data as required by the company.